refactor: Colors 등 prop MessageTo 컴포넌트로 이동
