+++ src/components/MessageTo/MessageTo.js
@@ -5,6 +5,28 @@ import InputPost from "./InputPost/InputPost";
 import styles from "./MessageTo.module.css";
 import MessageToToggleButton from "./MessageToToggleButton/MessageToToggleButton";
 
+/* eslint-disable */
+const backgroundColor = {
+  beige: "var(--Orange-20)",
+  purple: "var(--Purple-20)",
+  blue: "var(--Blue-20)",
+  green: "var(--Green-20)",
+};
+
+const backgroundImageURL = {
+  photo1: car,
+  photo2: cliff,
+  photo3: car,
+  photo4: cliff,
+};
+
+const Colors = [
+  backgroundColor.beige,
+  backgroundColor.purple,
+  backgroundColor.blue,
+  backgroundColor.green,
+];
+
 const MessageTo = () => {
   const [name, setName] = useState("");
   const [error, setError] = useState(false);
@@ -19,6 +41,13 @@ const MessageTo = () => {
       setError(true);
     }
   };
+
+  const recipientData = {
+    name: name || "",
+    backgroundColor: Colors || beige,
+    backgroundImageURL,
+  };
+
   return (
     <div className={styles.PostPageBody}>
       <InputPost
@@ -34,7 +63,11 @@ const MessageTo = () => {
             컬러를 선택하거나, 이미지를 선택할 수 있습니다.
           </div>
         </div>
-        <MessageToToggleButton />
+        <MessageToToggleButton
+          backgroundColor={backgroundColor}
+          backgroundImageURL={backgroundImageURL}
+          Colors={Colors}
+        />
       </div>
       <CreateButton name={name} />
     </div>
